alterar o widthFactor para float
changing the field to float it is possible to get smaller sizes, using 0.5 for example

+++ src/BarcodeGeneratorHTML.php
@@ -14,7 +14,7 @@ class BarcodeGeneratorHTML extends BarcodeGenerator
      * @param string $foregroundColor Foreground color for bar elements as '#333' or 'orange' for example (background is transparent).
      * @return string HTML code.
      */
-    public function getBarcode($barcode, $type, int $widthFactor = 2, int $height = 30, string $foregroundColor = 'black')
+    public function getBarcode($barcode, $type, float $widthFactor = 2, int $height = 30, string $foregroundColor = 'black')
     {
         $barcodeData = $this->getBarcodeData($barcode, $type);
